refactored what i could, full commit details to be added tomorrow
