+++ 4
@@ -1,0 +1,1 @@
+.PHONY: all clean fclean re test libftprintf
