+++ src/executor/redirection.h
@@ -16,9 +16,11 @@ typedef struct s_multi_pipes
 	int	previous[PIPE_BOTH_ENDS];
 }	t_multi_pipes;
 
+void	handle_stdin(int in_file, t_multi_pipes *pipes, int process);
+void	handle_stdout(int out_file, t_multi_pipes *pipes, int process,
+			int last_process);
 void	set_stdout(int new_std_out);
 void	set_stdin(int new_std_in);
-// void	get_file_name(char *dst_buffer, const char *src);
 void	previous_to_current_pipe(t_multi_pipes *pipes);
 void	redirect_in_and_output(t_multi_pipes *pipes,
 			int process, int num_of_processes, const char *cmd_string);
