+++ tests/pipe_tests/commands.txt
@@ -10,3 +10,5 @@ cat tests/pipe_tests/bigfile | wc -l | cat -e | cat -e
 < tests/pipe_tests/bigfile grep int < tests/pipe_tests/simple_file
 < tests/pipe_tests/bigfile grep int < tests/pipe_tests/simple_file | cat -e 
 <main.c<Makefile grep int
+<"Makefile" grep int
+<Makefile grep int
